rag_v2.1/requirements.txt
pandas>=2.0.0
numpy>=1.24.0
sentence-transformers>=2.2.0
scikit-learn>=1.3.0
torch>=2.0.0
# LangChain関連を削除（タグレス対応）
# langchain>=0.1.0
# langchain-anthropic>=0.0.1
# langchain-openai>=0.0.1
# langchain-google-genai>=0.0.1
google-generativeai>=0.3.0
google-cloud-aiplatform>=1.35.0
google-auth>=2.17.0
python-dotenv>=1.0.0
openpyxl>=3.1.2
xlsxwriter>=3.1.0
sudachipy>=0.6.8
sudachidict-core>=20230927
streamlit>=1.30.0
streamlit_elements==0.1.0
pymupdf4llm>=0.0.14
tqdm
# 新規追加: Vertex AI埋め込みモデル用
google-cloud-aiplatform[vertexai]>=1.35.0
# 新規追加: Azure Key Vault用
azure-identity>=1.15.0
azure-keyvault-secrets>=4.7.0
# 新規追加: メタデータ対応ベクトルDB用
chromadb>=1.0.15
# 新規追加: Gemini Embedding API用
google-cloud-aiplatform[vertexai]>=1.35.0